NMS-10398 rename CdpTopologyInfo... to TopologyInfo...
