+++ opennms-dao-api/src/main/java/org/opennms/netmgt/dao/api/TopologyInfoDao.java
@@ -26,27 +26,14 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.dao.hibernate;
+package org.opennms.netmgt.dao.api;
 
 import java.util.List;
 
-import org.opennms.netmgt.dao.api.CdpTopologyInfoDao;
 import org.opennms.netmgt.model.CdpLinkInfo;
 import org.opennms.netmgt.model.VertexInfo;
-import org.springframework.orm.hibernate3.support.HibernateDaoSupport;
 
-public class CdpTopologyInfoDaoHibernate extends HibernateDaoSupport implements CdpTopologyInfoDao {
-
-    @Override
-    public List<VertexInfo> getVertexInfos() {
-        return (List<VertexInfo>)getHibernateTemplate().find(
-                "select new org.opennms.netmgt.model.VertexInfo(n.id, n.type, n.sysObjectId, n.label, n.location) from org.opennms.netmgt.model.OnmsNode n");
-    }
-
-    @Override
-    public List<CdpLinkInfo> getCdpLinkInfo() {
-        return (List<CdpLinkInfo>)getHibernateTemplate().find(
-                "select new org.opennms.netmgt.model.CdpLinkInfo(l.id, l.node.id, l.cdpCacheIfIndex, " +
-                        "l.cdpInterfaceName, l.cdpCacheAddress, l.cdpCacheDeviceId, l.cdpCacheDevicePort) from org.opennms.netmgt.model.CdpLink l");
-    }
+public interface TopologyInfoDao {
+    List<VertexInfo> getVertexInfos();
+    List<CdpLinkInfo> getCdpLinkInfo();
 }
